Fix: buff Swashbuckler riposte feature
The Swashbuckler's Riposte feature triggers on a weak condition (on a
successful Deception check). By the rules, it's possible for anyone to
disorientate an enemy and therefore gain advantage in combat.

Instead, I'm replacing this with a tactical feature that encourages the
Swashbuckler to move in and out of close combat. In a way that
simulates dashing and jumping around.