Add @MediaQuery so the page can be responsive
I have integrated @mediaqueries so the website is responsive on all
devices. Some tweaks will have to be done on the mobile size website as
in adding functionality to the toggle button when in mobile size, and
the navbar to become full width and height when shown!

Also for easer readability the font-size of the page should be smaller
when in mobile size!

https://github.com/grislyeye/three-meet/pull/101

Co-authored-by: R.G. Wood <ric@grislyeye.com>