Fix: Swashbuckler sub-class not generic enough
The Swashbuckler class implies a fictional background. This goes against
the Three Meet design principle that classes provide purely combat
archetypes.

Re-named to a more generic Skirmisher sub-class. Players wanting to run
a swashbuckler might pick a Pirate or Noble background with this
sub-class. Players wanting to play a scout might take the Survival
proficiency with this sub-class.

https://github.com/grislyeye/three-meet/pull/94